Separated ButtonInfo into its own file and renamed to Button
